Modularized the push notification code so Home screen uses a hook instead from hooks folder
