+++ Project/SafeIllini/services/NotificationService.ts
@@ -1,0 +1,80 @@
+import { Platform } from "react-native";
+import * as Notifications from 'expo-notifications';
+import Constants from 'expo-constants';
+import * as Device from 'expo-device';
+
+// Code mostly from https://docs.expo.dev/push-notifications/push-notifications-setup/
+function handleRegistrationError(errorMessage: string) {
+  alert(errorMessage);
+  throw new Error(errorMessage);
+}
+
+class NotificationService {
+  async registerForPushNotificationsAsync() {
+    if (Platform.OS === 'android') {
+      await this.setAndroidChannel();
+    }
+
+    if (!Device.isDevice) {
+      handleRegistrationError('Must use physical device for push notifications');
+    }
+
+    const { status: existingStatus } = await Notifications.getPermissionsAsync();
+    let finalStatus = existingStatus;
+
+    if (existingStatus !== 'granted') {
+      const { status } = await Notifications.requestPermissionsAsync();
+      finalStatus = status;
+    }
+
+    if (finalStatus !== 'granted') {
+      handleRegistrationError('Permission not granted to get push token for push notification!');
+      return;
+    }
+    // projectID necessary for getExpoPushTokenAsync to work correctly
+    const projectId =
+      Constants?.expoConfig?.extra?.eas?.projectId ?? Constants?.easConfig?.projectId;
+    if (!projectId) {
+      handleRegistrationError('Project ID not found');
+    }
+
+    try {
+      const pushTokenString = (await Notifications.getExpoPushTokenAsync()).data;
+      console.log(pushTokenString);
+      return pushTokenString;
+    } catch (e: unknown) {
+      handleRegistrationError(`${e}`);
+    }
+  }
+
+  private async setAndroidChannel() {
+    await Notifications.setNotificationChannelAsync('default', {
+      name: 'default',
+      importance: Notifications.AndroidImportance.MAX,
+      vibrationPattern: [0, 250, 250, 250],
+      lightColor: '#FF231F7C',
+    });
+  }
+
+  async sendPushNotification(expoPushToken: string) {
+    const message = {
+      to: expoPushToken,
+      sound: 'default',
+      title: 'Original Title',
+      body: 'And here is the body!',
+      data: { someData: 'goes here' },
+    };
+
+    await fetch('https://exp.host/--/api/v2/push/send', {
+      method: 'POST',
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json',
+        'Accept-encoding': 'gzip, deflate',
+      },
+      body: JSON.stringify(message),
+    });
+  }
+};
+
+export default new NotificationService();
